Initial implementation of align(not working) and corner combining(working badly). I am probably mistreating dbscan and doing something else stupid, but it kinda seems to show promise and I am going insane while staring at these goddamn jumping lines and red dots.

+++ scitos_common/include/scitos_common/dbscan.hpp
@@ -46,7 +46,7 @@ std::vector<int> dbscan(const std::vector<T> &data,
     bool merged = cluster.size() > n;
     while (merged) {
       merged = false;
-      for (int j = 0; j < data.size(); j++) {
+      for (size_t j = 0; j < data.size(); j++) {
         if (lables.at(j) != 0 || cluster.find(j) != cluster.end()) {
           continue;
         }
@@ -115,7 +115,7 @@ std::vector<std::vector<T>> dbscan2(const std::vector<T> &data,
     bool merged = cluster.size() > n;
     while (merged) {
       merged = false;
-      for (int j = 0; j < data.size(); j++) {
+      for (size_t j = 0; j < data.size(); j++) {
         if (lables.at(j) != 0 || cluster.find(j) != cluster.end()) {
           continue;
         }
